requirements.txt
streamlit
pandas
hyperliquid-python-sdk
plotly
streamlit-autorefresh